refactor: rename utility components and update import paths for consistency
