+++ src/components/utils/ui/Create.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { Button } from "../ui/button";
+import { Button } from "../../ui/button";
 
 interface CreateButtonProps {
   label: string;
